RE RE RE REDID ALL OF THE THINGS!
testTags.html; Went back to a Form design closer to the original in setup, but updated with the th:with fieldName trick. This uses a new shared formatting to fix the DB and Attributes not being aligned 1-1.

'/search/'; Added and updated how the form is read, should now just be 1 line of code "queryService.buildFromForm(form);"

CategoryTagMapService.java; Remade the tagDataFiller to add the new Formatting

Search Still not entirely done tho